--- productFolder/product.js
@@ -1,15 +1,0 @@
-"use strict";
-module.exports = function(sequelize, DataTypes) {
-  var Product = sequelize.define(
-    "Product",
-    {
-      name: DataTypes.STRING,
-      sku: DataTypes.STRING,
-      description: DataTypes.STRING,
-      price: DataTypes.FLOAT,
-      categoryId: DataTypes.INTEGER
-    },
-    { Product.hasOne(models.Category);}
-  );
-  return Product;
-};
