+++ seeds/seq/20170809205430-categories.js
@@ -1,25 +1,19 @@
-'use strict';
+"use strict";
 
 let cats = ["weapons", "balloons", "chairs", "onions"];
 
 module.exports = {
-  up: function (queryInterface, Sequelize) {
-    /*
-      Add altering commands here.
-      Return a promise to correctly handle asynchronicity.
-
-      Example:
-      return queryInterface.bulkInsert('Person', [{
-        name: 'John Doe',
-        isBetaMember: false
-      }], {});
-    */
-
-   return queryInterface.bulkInsert("Categories", cats);
-
+  up: function(queryInterface, Sequelize) {
+    let categories = [];
+    cats.forEach(i => {
+      categories.push({
+        name: cats[i]
+      });
+    });
+    return queryInterface.bulkInsert("Categories", categories);
   },
 
-  down: function (queryInterface, Sequelize) {
+  down: function(queryInterface, Sequelize) {
     /*
       Add reverting commands here.
       Return a promise to correctly handle asynchronicity.
